+++ router/subscribe_to_pay_via_details.js
@@ -9,19 +9,28 @@ const randomId = () => randomBytes(32).toString('hex');
 
   Requires LND built with `routerrpc` build tag
 
+  Specifying `features` is not supported on LND 0.8.2 and below
   Specifying `max_fee_mtokens`/`mtokens` is not supported in LND 0.8.2 or below
+  Specifying `messages` is not supported on LND 0.8.2 and below
 
   `incoming_peer` is not supported on LND 0.8.2 and below
 
   {
     [cltv_delta]: <Final CLTV Delta Number>
     destination: <Destination Public Key String>
+    [features]: [{
+      bit: <Feature Bit Number>
+    }]
     [id]: <Payment Request Hash Hex String>
     [incoming_peer]: <Pay Through Specific Final Hop Public Key Hex String>
     lnd: <Authenticated LND gRPC API Object>
     [max_fee]: <Maximum Fee Tokens To Pay Number>
     [max_fee_mtokens]: <Maximum Fee Millitokens to Pay String>
     [max_timeout_height]: <Maximum Height of Payment Timeout Number>
+    [messages]: [{
+      type: <Message Type Number String>
+      value: <Message Raw Value Hex Encoded String>
+    }]
     [mtokens]: <Millitokens to Pay String>
     [outgoing_channel]: <Pay Out of Outgoing Channel Id String>
     [pathfinding_timeout]: <Time to Spend Finding a Route Milliseconds Number>
@@ -106,12 +115,14 @@ module.exports = args => {
   return subscribeToPay({
     cltv_delta: args.cltv_delta || defaultCltvDelta,
     destination: args.destination,
+    features: args.features,
     id: args.id || randomId(),
     incoming_peer: args.incoming_peer,
     lnd: args.lnd,
     max_fee: args.max_fee,
     max_fee_mtokens: args.max_fee_mtokens,
     max_timeout_height: args.max_timeout_height,
+    messages: args.messages,
     mtokens: args.mtokens,
     outgoing_channel: args.outgoing_channel,
     pathfinding_timeout: args.pathfinding_timeout,