+++ router/subscribe_to_pay_via_request.js
@@ -5,6 +5,7 @@ const subscribeToPay = require('./subscribe_to_pay');
   Requires LND built with `routerrpc` build tag
 
   Specifying `max_fee_mtokens`/`mtokens` is not supported in LND 0.8.2 or below
+  Specifying `messages` is not supported on LND 0.8.2 and below
 
   `incoming_peer` is not supported on LND 0.8.2 and below
 
@@ -14,6 +15,10 @@ const subscribeToPay = require('./subscribe_to_pay');
     [max_fee]: <Maximum Fee Tokens To Pay Number>
     [max_fee_mtokens]: <Maximum Fee Millitokens to Pay String>
     [max_timeout_height]: <Maximum Height of Payment Timeout Number>
+    [messages]: [{
+      type: <Message Type Number String>
+      value: <Message Raw Value Hex Encoded String>
+    }]
     [mtokens]: <Millitokens to Pay String>
     [outgoing_channel]: <Pay Out of Outgoing Channel Id String>
     [pathfinding_timeout]: <Time to Spend Finding a Route Milliseconds Number>
@@ -92,6 +97,7 @@ module.exports = args => {
     max_fee: args.max_fee,
     max_fee_mtokens: args.max_fee_mtokens,
     max_timeout_height: args.max_timeout_height,
+    messages: args.messages,
     mtokens: args.mtokens,
     outgoing_channel: args.outgoing_channel,
     pathfinding_timeout: args.pathfinding_timeout,