+++ test/invoices/test_htlc_as_payment.js
@@ -2,79 +2,74 @@ const {test} = require('tap');
 
 const {htlcAsPayment} = require('./../../invoices');
 
+const makeHtlc = overrides => {
+  const htlc = {
+    accept_height: 1,
+    accept_time: '1',
+    amt_msat: '1',
+    chan_id: '1',
+    custom_records: {
+      '\u0010\u0000\u0000\u0000\u0000\u0000\u0000\u0000': Buffer.alloc(1, 1),
+    },
+    expiry_height: 1,
+    htlc_index: '1',
+    resolve_time: '1',
+    state: 'SETTLED',
+  };
+
+  Object.keys(overrides).forEach(k => htlc[k] = overrides[k]);
+
+  return htlc;
+};
+
 const tests = [
   {
-    args: {},
+    args: makeHtlc({accept_height: undefined}),
     description: 'Accept height is needed',
     error: 'ExpectedAcceptHeightInResponseHtlc',
   },
   {
-    args: {accept_height: 1},
+    args: makeHtlc({accept_time: undefined}),
     description: 'Accept time is needed',
     error: 'ExpectedAcceptTimeInResponseHtlc',
   },
   {
-    args: {accept_height: 1, accept_time: 1},
+    args: makeHtlc({amt_msat: undefined}),
     description: 'Amount is needed',
     error: 'ExpectedPaymentAmountInResponseHtlc',
   },
   {
-    args: {accept_height: 1, accept_time: '1', amt_msat: '1'},
+    args: makeHtlc({chan_id: undefined}),
     description: 'Chan id is needed',
     error: 'ExpectedChannelIdInResponseHtlc',
   },
   {
-    args: {accept_height: 1, accept_time: '1', amt_msat: '1', chan_id: '1'},
+    args: makeHtlc({custom_records: undefined}),
+    description: 'Custom records are needed',
+    error: 'ExpectedCustomRecordsInResponseHtlc',
+  },
+  {
+    args: makeHtlc({expiry_height: undefined}),
     description: 'Expiry height is needed',
     error: 'ExpectedHtlcExpiryHeightInResponseHtlc',
   },
   {
-    args: {
-      accept_height: 1,
-      accept_time: '1',
-      amt_msat: '1',
-      chan_id: '1',
-      expiry_height: 1,
-    },
+    args: makeHtlc({htlc_index: undefined}),
     description: 'HTLC index is required',
     error: 'ExpectedHtlcIndexInResponseHtlc',
   },
   {
-    args: {
-      accept_height: 1,
-      accept_time: '1',
-      amt_msat: '1',
-      chan_id: '1',
-      expiry_height: 1,
-      htlc_index: '1',
-    },
+    args: makeHtlc({resolve_time: undefined}),
     description: 'Resolve time is required',
     error: 'ExpectedResolveTimeInResponseHtlc',
   },
   {
-    args: {
-      accept_height: 1,
-      accept_time: '1',
-      amt_msat: '1',
-      chan_id: '1',
-      expiry_height: 1,
-      htlc_index: '1',
-      resolve_time: '1',
-    },
-    description: 'Resolve time is required',
+    args: makeHtlc({state: undefined}),
+    description: 'Htlc state is required',
     error: 'ExpectedHtlcStateInResponseHtlc',
   },
   {
-    args: {
-      accept_height: 1,
-      accept_time: '1',
-      amt_msat: '1',
-      chan_id: '1',
-      expiry_height: 1,
-      htlc_index: '1',
-      resolve_time: '1',
-      state: 'SETTLED',
-    },
+    args: makeHtlc({mpp_total_amt_msat: '0'}),
     description: 'HTLC mapped to payment',
     expected: {
       canceled_at: null,
@@ -85,6 +80,7 @@ const tests = [
       is_canceled: false,
       is_confirmed: true,
       is_held: false,
+      messages: [{type: '16', value: '01'}],
       mtokens: '1',
       pending_index: undefined,
       timeout: 1,
@@ -93,16 +89,7 @@ const tests = [
     },
   },
   {
-    args: {
-      accept_height: 1,
-      accept_time: '1',
-      amt_msat: '1',
-      chan_id: '1',
-      expiry_height: 1,
-      htlc_index: '1',
-      resolve_time: '1',
-      state: 'CANCELED',
-    },
+    args: makeHtlc({amt_msat: '0', state: 'CANCELED'}),
     description: 'Canceled HTLC mapped to payment',
     expected: {
       canceled_at: '1970-01-01T00:00:01.000Z',
@@ -113,7 +100,8 @@ const tests = [
       is_canceled: true,
       is_confirmed: false,
       is_held: false,
-      mtokens: '1',
+      messages: [{type: '16', value: '01'}],
+      mtokens: '0',
       pending_index: undefined,
       timeout: 1,
       tokens: 0,
@@ -121,17 +109,8 @@ const tests = [
     },
   },
   {
-    args: {
-      accept_height: 1,
-      accept_time: '1',
-      amt_msat: '1',
-      chan_id: '1',
-      expiry_height: 1,
-      htlc_index: '1',
-      resolve_time: '0',
-      state: 'ACCEPTED',
-    },
-    description: 'Canceled HTLC mapped to payment',
+    args: makeHtlc({mpp_total_amt_msat: '1', state: 'ACCEPTED'}),
+    description: 'Accepted HTLC mapped to payment',
     expected: {
       canceled_at: null,
       confirmed_at: null,
@@ -141,11 +120,12 @@ const tests = [
       is_canceled: false,
       is_confirmed: false,
       is_held: true,
+      messages: [{type: '16', value: '01'}],
       mtokens: '1',
       pending_index: 1,
       timeout: 1,
       tokens: 0,
-      total_mtokens: undefined,
+      total_mtokens: '1',
     },
   },
 ];