+++ test/invoicesrpc-integration/test_subscribe_settle_invoice.js
@@ -1,6 +1,7 @@
 const {createHash} = require('crypto');
 const {randomBytes} = require('crypto');
 
+const asyncRetry = require('async/retry');
 const {test} = require('tap');
 
 const {createCluster} = require('./../macros');
@@ -9,7 +10,7 @@ const {delay} = require('./../macros');
 const {getInvoice} = require('./../../');
 const {getInvoices} = require('./../../');
 const {openChannel} = require('./../../');
-const {pay} = require('./../../');
+const {payViaPaymentRequest} = require('./../../');
 const {settleHodlInvoice} = require('./../../');
 const {subscribeToInvoice} = require('./../../');
 const {waitForChannel} = require('./../macros');
@@ -19,6 +20,10 @@ const channelCapacityTokens = 1e6;
 const confirmationCount = 6;
 const defaultFee = 1e3;
 const defaultVout = 0;
+const interval = retryCount => 50 * Math.pow(2, retryCount);
+const times = 15;
+const tlvType = '67890';
+const tlvValue = '00';
 const tokens = 100;
 
 // Subscribe to a settled invoice should return invoice settled event
@@ -32,8 +37,8 @@ test(`Subscribe to settled invoice`, async ({deepIs, end, equal}) => {
     lnd,
     chain_fee_tokens_per_vbyte: defaultFee,
     local_tokens: channelCapacityTokens,
-    partner_public_key: cluster.target_node_public_key,
-    socket: `${cluster.target.listen_ip}:${cluster.target.listen_port}`,
+    partner_public_key: cluster.target.public_key,
+    socket: cluster.target.socket,
   });
 
   await waitForPendingChannel({
@@ -70,6 +75,15 @@ test(`Subscribe to settled invoice`, async ({deepIs, end, equal}) => {
   equal(!!currentInvoice.is_confirmed, false, 'Invoice is not confirmed yet');
 
   setTimeout(async () => {
+    // Wait for the invoice to be held
+    await asyncRetry({interval, times}, async () => {
+      if (!currentInvoice.is_held) {
+        throw new Error('ExpectedInvoiceHeld');
+      }
+
+      return;
+    });
+
     equal(!!currentInvoice.is_held, true, 'Invoice is not held yet');
     equal(!!currentInvoice.is_canceled, false, 'Invoice is not canceled yet');
     equal(!!currentInvoice.is_confirmed, false, 'Invoice is confirmed');
@@ -79,7 +93,27 @@ test(`Subscribe to settled invoice`, async ({deepIs, end, equal}) => {
       secret: secret.toString('hex'),
     });
 
-    await delay(1000);
+    // Wait for the invoice to be confirmed
+    await asyncRetry({interval, times}, async () => {
+      if (!currentInvoice.is_confirmed) {
+        throw new Error('ExpectedInvoiceConfirmed');
+      }
+
+      return;
+    });
+
+    const {payments} = currentInvoice;
+
+    if (!!payments.length) {
+      const [payment] = payments;
+
+      if (!!payment.messages.length) {
+        const [{type, value}] = payment.messages;
+
+        equal(type, tlvType, 'Payment message TLV type returned');
+        equal(value, tlvValue, 'Payment message TLV value returned');
+      }
+    }
 
     equal(!!currentInvoice.is_held, false, 'Invoice is not held yet');
     equal(!!currentInvoice.is_canceled, false, 'Invoice is not canceled yet');
@@ -92,7 +126,11 @@ test(`Subscribe to settled invoice`, async ({deepIs, end, equal}) => {
   },
   1000);
 
-  const paid = await pay({lnd, request: invoice.request});
+  const paid = await payViaPaymentRequest({
+    lnd,
+    messages: [{type: tlvType, value: tlvValue}],
+    request: invoice.request,
+  });
 
   equal(paid.secret, secret.toString('hex'), 'Paying reveals the HTLC secret');